requirements.txt
django-registration==1.0
Pillow==2.9.0
wheel==0.24.0
